+++ packages/http-client/RELEASES.md
@@ -6,6 +6,8 @@
   - Types implementing the method `RequestHandler.handleAuthentication()` now throw an `Error` rather than returning `null` if they do not support handling an HTTP 401 response. Callers can still use `canHandleAuthentication()` to determine if this handling is supported or not.
   - Types using `any` have been scoped to more specific types.
 - Following TypeScript's naming conventions, exported interfaces no longer begin with the prefix `I-`.
+- Delete the `IHttpClientResponse` interface in favor of the `HttpClientResponse` class.
+- Delete the `IHeaders` interface in favor of `http.OutgoingHttpHeaders`.
 - The source code of the package was moved to build with [actions/toolkit](https://github.com/actions/toolkit).
 
 ## 1.0.11